Separated pages, components and middleware.
