+++ src/pages/SignUp.js
@@ -15,7 +15,7 @@ import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import Alert from '@mui/material/Alert';
-
+import API_URL from '../config/config'
 
 const theme = createTheme();
 
@@ -78,7 +78,7 @@ const SignUp = () => {
       setError(true);
       setErrorMessage("Please fill all the fields");
     } else {
-      const response = await fetch("http://localhost:1212/signup", {
+      const response = await fetch(`${API_URL}/signup`, {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
